requirements.txt
aiohttp-cors==0.5.2
aiohttp==2.0.7
itsdangerous==0.24
jsonschema==2.6.0
PyYAML==3.12

-e git+https://github.com/IL2HorusTeam/il2fb-commons.git#egg=il2fb-commons
-e git+https://github.com/IL2HorusTeam/il2fb-difficulty.git#egg=il2fb-difficulty
-e git+https://github.com/IL2HorusTeam/il2fb-events-parser.git#egg=il2fb-events-parser
-e git+https://github.com/IL2HorusTeam/il2fb-github-bug-reporter.git#egg=il2fb-github-bug-reporter
